// VCG/src/interfaces/sendStatusEmailTypes.tsx
export interface ISentEmailForm {
    eventId:  string
    eventStatus: string
    emailSubject:  string
    emailContent: string
   
}

export interface IFormStatus {
    message: string
    type: string
}



export interface ISentEmailForm {
    eventId:  string
    eventStatus: string
    emailSubject:  string
    emailContent: string
}